requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
pandas
lxml
html5lib
matplotlib
line-bot-sdk
python-dotenv